Update question 8

+++ Problem Set/RPS_Mixed/RPS_Mixed/RPSController.m
@@ -15,7 +15,7 @@
     _game = game;
 }
 
--(void)throwDown:(int) playersMove {
+-(void)throwDown:(NSInteger) playersMove {
     
     // Here the RPSTurn class generates the opponent's move
     RPSTurn *playersTurn = [[RPSTurn alloc]initWithMove:playersMove];
